Move main API module to /src
