--- api/Dockerfile
@@ -1,10 +1,0 @@
-FROM python:3.11-slim
-
-COPY . /app/
-WORKDIR /app
-
-RUN pip3 install -r requirements.txt
-
-EXPOSE 8000
-
-CMD ["uvicorn", "main:app", "--host=0.0.0.0", "--reload"]
